fix: avatar collision
